+++ frameworks/react.ts
@@ -1,0 +1,5 @@
+import {reactMountPlugin, getStaticInfoPlugin} from "../plugins"
+
+export default {
+    plugins: [reactMountPlugin(), getStaticInfoPlugin()],
+}
