+++ frameworks/svelte.ts
@@ -1,0 +1,5 @@
+import {svelteMountPlugin} from "../plugins"
+
+export default {
+    plugins: [svelteMountPlugin()],
+}
