+++ globals.d.ts
@@ -4,3 +4,10 @@ declare module 'getFrameworkViews' {
         path: string;
     }>
 }
+
+// declare process.env with the framework var
+declare namespace NodeJS {
+    interface ProcessEnv {
+        FRAMEWORK: string;
+    }
+}
