+++ tsconfig.json
@@ -26,10 +26,11 @@
         // "useDefineForClassFields": true,                  /* Emit ECMAScript-standard-compliant class fields. */
 
         /* Modules */
-        "module": "commonjs",
+        "module": "ESNext",
         /* Specify what module code is generated. */
         // "rootDir": "./",                                  /* Specify the root folder within your source files. */
-        // "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from a given module specifier. */
+        "moduleResolution": "NodeNext",
+        /* Specify how TypeScript looks up a file from a given module specifier. */
         // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
         // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
@@ -104,5 +105,5 @@
         // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
         "skipLibCheck": true /* Skip type checking all .d.ts files. */
     },
-    "include": ["globals.d.ts"]
+    "include": ["globals.d.ts", "rollup.config.ts"]
 }
